+++ src/components/rightSidePanel/subgraph/SubgraphNodeWidget.vue
